Implement prepend instrumentation for Typhoeus (#612)
* update typhoeus instrumentation
and add prepend

* more DRY implementation; changed naming to match other instrumentation conventions
